+++ Extension/src/pages/filtering-log/stores/LogStore.ts
@@ -512,6 +512,10 @@ class LogStore {
             preserveLogEnabled,
         } = await messenger.getFilteringLogData();
 
+        // telemetry
+        const anonymizedUsageDataAllowed = settings.values[settings.names.AllowAnonymizedUsageData];
+        this.rootStore.telemetryStore.setIsAnonymizedUsageDataAllowed(anonymizedUsageDataAllowed);
+
         runInAction(() => {
             this.filtersMetadata = filtersMetadata;
             this.settings = settings;