+++ tools/bundle-size/check.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2015-2025 Adguard Software Ltd.
+ * Copyright (c) 2015-2026 Adguard Software Ltd.
  *
  * @file
  * This file is part of AdGuard Browser Extension (https://github.com/AdguardTeam/AdguardBrowserExtension).
@@ -41,6 +41,8 @@ import {
     Browser,
     isValidBrowserTarget,
     isValidBuildEnv,
+    MV3_BROWSERS,
+    type Mv3Browser,
 } from '../constants';
 import { getBrowserConf } from '../bundle/helpers';
 
@@ -266,21 +268,22 @@ function compareBuildSizes(
 }
 
 /**
- * Check the size of the Chrome MV3 bundle.
+ * Check the size of the MV3 bundle.
  *
  * @param buildType Build environment (beta, release, etc.).
+ * @param targetMv3Browser Target MV3 browser.
  *
  * @returns True if the size exceeds the limit, else false.
  */
-async function checkChromeMv3BundleSize(buildType: BuildTargetEnv): Promise<boolean> {
-    console.log('\n\nChecking Chrome MV3 bundle size...');
+async function checkMv3BundleSize(buildType: BuildTargetEnv, targetMv3Browser: Mv3Browser): Promise<boolean> {
+    console.log(`\n\nChecking "${targetMv3Browser}" bundle size...`);
 
     try {
         // Get current build stats for this target
-        const currentStats = await getCurrentBuildStats(buildType, Browser.ChromeMv3);
+        const currentStats = await getCurrentBuildStats(buildType, targetMv3Browser);
 
         const mv3Size = currentStats.stats.zip;
-        const zipArchiveName = `${getBrowserConf(Browser.ChromeMv3).zipName}${ZIP_EXTENSION}`;
+        const zipArchiveName = `${getBrowserConf(targetMv3Browser).zipName}${ZIP_EXTENSION}`;
 
         if (mv3Size && mv3Size > MAX_MV3_SIZE_BYTES) {
             console.error(`${zipArchiveName}: ${(mv3Size / (1024 * 1024)).toFixed(2)}MB - Exceeds maximum allowed size of 30MB! ❌`);
@@ -288,11 +291,11 @@ async function checkChromeMv3BundleSize(buildType: BuildTargetEnv): Promise<bool
             return true;
         }
 
-        console.log('✅ Chrome MV3 bundle size is ok!');
+        console.log(`✅ "${targetMv3Browser}" bundle size is ok!`);
 
         return false;
     } catch (e) {
-        console.error(`Error checking Chrome MV3 bundle size: ${e}`);
+        console.error(`Error checking "${targetMv3Browser}" bundle size: ${e}`);
 
         return true;
     }
@@ -467,11 +470,16 @@ async function checkBundleSizes({ buildEnv, targetBrowser, threshold }: CheckBun
         }
     }
 
-    // Check max size only for chrome-mv3 target, because we pack a lot
-    // of filters data inside this target.
-    let hasChromeMv3SizeIssues = false;
-    if (targets.includes(Browser.ChromeMv3)) {
-        hasChromeMv3SizeIssues = await checkChromeMv3BundleSize(buildEnv);
+    // Check max size for mv3 targets, because we pack a lot of filters data inside this target
+    let hasMv3SizeIssues = false;
+    for (const mv3Target of MV3_BROWSERS) {
+        if (!targets.includes(mv3Target)) {
+            continue;
+        }
+
+        // Note: We need to check every MV3 target, thats why it's not inside of the boolean expression
+        const isTargetHasMv3SizeIssues = await checkMv3BundleSize(buildEnv, mv3Target);
+        hasMv3SizeIssues = hasMv3SizeIssues || isTargetHasMv3SizeIssues;
     }
 
     let hasFirefoxAmoSizeIssues = false;
@@ -495,7 +503,7 @@ async function checkBundleSizes({ buildEnv, targetBrowser, threshold }: CheckBun
     // Exit with error if there are issues in any target
     if (
         hasSizeIssues
-        || hasChromeMv3SizeIssues
+        || hasMv3SizeIssues
         || hasFirefoxAmoSizeIssues
         || hasDuplicates
         || hasFirefoxJsIssues